Move UseCase classes to usecase.base package
