+++ domain/src/main/java/dev/mkeeda/spaceship/domain/usecase/ShowSpacePostContext.kt
@@ -3,9 +3,10 @@ package dev.mkeeda.spaceship.domain.usecase
 import dev.mkeeda.spaceship.data.PostingLocation
 import dev.mkeeda.spaceship.data.TimelinePostDetail
 import dev.mkeeda.spaceship.domain.repository.ThreadRepository
+import dev.mkeeda.spaceship.domain.usecase.base.UseCase
+import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import javax.inject.Inject
 
 class ShowSpacePostContext @Inject constructor(
     private val threadRepository: ThreadRepository
